@@ -9,26 +9,26 @@
     "format": "prettier --ignore-path .gitignore --write src/"
   },
   "devDependencies": {
-    "@sveltejs/adapter-static": "3.0.6",
-    "@sveltejs/kit": "2.12.2",
+    "@sveltejs/adapter-static": "3.0.8",
+    "@sveltejs/kit": "2.15.1",
     "@sveltejs/vite-plugin-svelte": "^5.0.3",
     "@svgdotjs/svg.js": "3.2.4",
     "clipboard": "2.0.11",
     "eslint": "8.57.0",
     "eslint-config-prettier": "9.1.0",
     "eslint-plugin-svelte": "^2.46.1",
-    "highlight.js": "11.11.0",
+    "highlight.js": "11.11.1",
     "normalize.css": "^8.0.1",
     "postcss": "8.4.49",
     "postcss-import": "16.1.0",
     "postcss-load-config": "6.0.1",
-    "postcss-preset-env": "10.1.2",
+    "postcss-preset-env": "10.1.3",
     "postcss-reporter": "^7.1.0",
     "prettier": "3.4.2",
     "prettier-plugin-svelte": "3.3.2",
-    "svelte": "5.14.4",
+    "svelte": "5.16.0",
     "svelte-preprocess": "6.0.3",
-    "vite": "6.0.4"
+    "vite": "6.0.6"
   },
   "type": "module"
 }
